do exec

+++ exercises/structs/structs3.rs
@@ -7,8 +7,6 @@
 // Execute `rustlings hint structs3` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 #[derive(Debug)]
 struct Package {
     sender_country: String,
@@ -42,8 +40,10 @@ impl Package {
     // TODO: Add the correct return type to the function signature.
     fn get_fees(&self, cents_per_gram: u32)->u32
     {
+
+
         // TODO: Calculate the package's fees.
-        cents_per_gram * self.weight_in_grams
+        cents_per_gram * (self.weight_in_grams as u32)
     }
 }
